cf_v2_src/requirements.txt
# Function dependencies, for example:
# package>=version
pandas-datareader
SQLAlchemy
psycopg2
SQLAlchemy-Utils
google-cloud-pubsub
yfinance